+++ fuocore/__init__.py
@@ -1,6 +1,5 @@
 import logging
 import logging.config
-# from logging import NullHandler
 
 from fuocore.models import ModelType  # noqa
 from fuocore.player import (
@@ -10,7 +9,7 @@ from fuocore.player import (
     Playlist,
 )  # noqa
 from fuocore.live_lyric import LiveLyric  # noqa
-from .library import Library
+from .library import Library  # noqa
 
 
 __all__ = [
@@ -19,48 +18,4 @@ __all__ = [
     'PlaybackMode',
     'Playlist',
     'LiveLyric',
-    'ModelType',
 ]
-
-
-dict_config = {
-    'version': 1,
-    'disable_existing_loggers': False,
-    'formatters': {
-        'standard': {
-            'format': "[%(levelname)s] "
-                      "[%(module)s %(funcName)s %(lineno)d] "
-                      ": %(message)s",
-        },
-    },
-    'handlers': {
-        'debug': {
-            'formatter': 'standard',
-            'level': 'DEBUG',
-            'class': 'logging.StreamHandler',
-        },
-        'release': {
-            'formatter': 'standard',
-            'level': 'INFO',
-            'class': 'logging.StreamHandler',
-        }
-    },
-    'loggers': {
-        'fuocore': {
-            'handlers': ['debug'],
-            'level': logging.DEBUG,
-            'propagate': True
-        }
-    }
-}
-
-
-def setup_logger(debug=True):
-    if debug:
-        logging.config.dictConfig(dict_config)
-    else:
-        dict_config['loggers']['fuocore']['handlers'] = ['release']
-        logging.config.dictConfig(dict_config)
-
-
-# logging.getLogger(__name__).addHandler(NullHandler())
