+++ fuocore/app.py
@@ -1,6 +1,8 @@
 """
 fuocore.app
 ~~~~~~~~~~~
+
+It provides a CliMixin to make integration easilly.
 """
 
 import asyncio
@@ -10,9 +12,6 @@ from collections import defaultdict
 from urllib.parse import urlparse
 
 from fuocore.aio_tcp_server import TcpServer
-from fuocore.pubsub import run as run_pubsub
-
-from fuocore import setup_logger
 from fuocore import MpvPlayer
 from fuocore import LiveLyric
 from fuocore.furi import parse_furi
@@ -33,22 +32,22 @@ class LiveLyricPublisher(object):
         self.gateway.publish(sentence + '\n', self.topic)
 
 
-class CliMixin(object):
-    def __init__(self, pubsub_gateway):
+class CliAppMixin(object):
+    """
+    FIXME: Subclass must call init to make this mixin
+    work properly, which seems to be little bit strange. But
+    this mixin helps avoid duplicate code temporarily.
+    """
+    def __init__(self):
         live_lyric = LiveLyric()
-        live_lyric_publisher = LiveLyricPublisher(pubsub_gateway)
+        live_lyric_publisher = LiveLyricPublisher(self.pubsub_gateway)
+
+        self.live_lyric = live_lyric
+        self._live_lyric_publisher = live_lyric_publisher
+
         live_lyric.sentence_changed.connect(live_lyric_publisher.publish)
         self.player.position_changed.connect(live_lyric.on_position_changed)
         self.playlist.song_changed.connect(live_lyric.on_song_changed)
-        self._live_lyric = live_lyric
-        self._live_lyric_publisher = live_lyric_publisher
-
-    def list_providers(self):
-        return self.library.list()
-
-    def get_provider(self, name):
-        # XXX: maybe this interface should removed
-        return self.library.get(name)
 
 
 async def handle(conn, addr, app, live_lyric):
@@ -72,10 +71,10 @@ async def handle(conn, addr, app, live_lyric):
         event_loop.sock_sendall(conn, bytes(msg, 'utf-8'))
 
 
-async def run(app, live_lyric, *args, **kwargs):
+async def run_server(app, live_lyric, *args, **kwargs):
     port = 23333
     host = '0.0.0.0'
     event_loop = asyncio.get_event_loop()
     event_loop.create_task(
         TcpServer(host, port, handle_func=handle).run(app, live_lyric))
-    logger.info('Fuo daemon running at {}:{}'.format(host, port))
+    logger.info('Fuo daemon run in {}:{}'.format(host, port))
